+++ lib/entities/entities.dart
@@ -1,0 +1,4 @@
+export 'address_component.dart';
+export 'auto_complete_item.dart';
+export 'location_result.dart';
+export 'near_by_place.dart';
