+++ lib/widgets/select_place_action.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class SelectPlaceAction extends StatelessWidget {
+  final String locationName;
+  final VoidCallback onTap;
+
+  SelectPlaceAction(this.locationName, this.onTap);
+
+  @override
+  Widget build(BuildContext context) {
+    return Material(
+      child: InkWell(
+        onTap: onTap,
+        child: Container(
+          padding: EdgeInsets.symmetric(horizontal: 24, vertical: 16),
+          child: Row(
+            children: <Widget>[
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: <Widget>[
+                    Text(locationName, style: TextStyle(fontSize: 16)),
+                    Text("Tap to select this location", style: TextStyle(color: Colors.grey, fontSize: 15)),
+                  ],
+                ),
+              ),
+              Icon(Icons.arrow_forward)
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
